refactor: made the /leaderboard route fully SSR except searchbar
