+++ app/leaderboard/_components/LeaderboardWrapper.tsx
@@ -1,20 +1,15 @@
 import { getLeaderboardData } from "../../api/leaderboard/functions";
 import { parseDateRangeSearchParam } from "@/lib/utils";
-import Leaderboard, { LeaderboardSortKey } from "../Leaderboard";
+import Leaderboard from "./Leaderboard";
+import { LeaderboardPageProps } from "../page";
 
 export default async function LeaderboardWrapper({
   searchParams,
-}: {
-  searchParams: {
-    between?: string; // <start-date>...<end-date>
-    sortBy?: LeaderboardSortKey | `-${LeaderboardSortKey}`;
-    roles?: string; // typeof subsetOf("core", "intern", "operations", "contributor").join(',')
-  };
-}) {
+}: LeaderboardPageProps) {
   const data = await getLeaderboardData(
     parseDateRangeSearchParam(searchParams.between),
     searchParams.sortBy ?? "-points",
   );
 
-  return <Leaderboard data={data} />;
+  return <Leaderboard data={data} searchParams={searchParams} />;
 }
