+++ app/leaderboard/page.tsx
@@ -1,22 +1,30 @@
 import { Suspense } from "react";
-import Leaderboard, { LeaderboardSortKey } from "./Leaderboard";
+import Leaderboard, { LeaderboardSortKey } from "./_components/Leaderboard";
 import LeaderboardWrapper from "./_components/LeaderboardWrapper";
 import LoadingText from "@/components/LoadingText";
+import Searchbar from "./_components/Searchbar";
 
-type PageProps = {
+export type LeaderboardPageProps = {
   searchParams: {
+    search?: string;
     between?: string; // <start-date>...<end-date>
     sortBy?: LeaderboardSortKey | `-${LeaderboardSortKey}`;
     roles?: string; // typeof subsetOf("core", "intern", "operations", "contributor").join(',')
   };
 };
 
-export default async function LeaderboardPage({ searchParams }: PageProps) {
+export default async function LeaderboardPage({
+  searchParams,
+}: LeaderboardPageProps) {
   return (
     <section className="bg-background text-foreground border-t dark:border-gray-700 border-gray-300">
-      <Suspense fallback={<LoadingText text="Ranking the contributors" />}>
-        <LeaderboardWrapper searchParams={searchParams} />
-      </Suspense>
+      <div className="max-w-6xl mx-auto">
+        <Searchbar />
+        <div className="border-gray-600 mx-4 xl:mx-0"></div>
+        <Suspense fallback={<LoadingText text="Ranking the contributors" />}>
+          <LeaderboardWrapper searchParams={searchParams} />
+        </Suspense>
+      </div>
     </section>
   );
 }
